+++ src/public.js
@@ -107,6 +107,17 @@ Grid.defaults = {
      **/
     keepSelection: false,
 
+    /**
+     * Defines whether the columns which are filtered or sorted should be highlighted or not.
+     *
+     * @property highlightColumns
+     * @type Boolean
+     * @default false
+     * @for defaults
+     * @since 1.2.0
+     **/
+    highlightColumns: false,
+
     highlightRows: false, // highlights new rows (find the page of the first new row)
     sorting: true,
     multiSort: false,
@@ -199,6 +210,18 @@ Grid.defaults = {
      **/
     css: {
         actions: "actions btn-group", // must be a unique class name or constellation of class names within the header and footer
+
+        /**
+         * CSS class to highlight active parts like sorted or filtered columns.
+         *
+         * @property active
+         * @type String
+         * @default "active"
+         * @for css
+         * @since 1.2.0
+         **/
+        active: "active",
+
         center: "text-center",
         columnHeaderAnchor: "column-header-anchor", // must be a unique class name or constellation of class names within the column header cell
         columnHeaderText: "text",
@@ -320,7 +343,7 @@ Grid.prototype.append = function(rows)
 {
     if (this.options.ajax)
     {
-        // todo: implement ajax DELETE
+        // todo: implement ajax POST
     }
     else
     {
@@ -384,8 +407,7 @@ Grid.prototype.destroy = function()
     {
         this.footer.remove();
     }
-    // todo: find a better and shorter way to remove events, data and the table itself!
-    this.element.before(this.original).off(namespace).removeData(namespace).remove();
+    this.element.before(this.original).remove();
 
     return this;
 };