+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed table column
 - Sorted header column CSS class
 - Improved destroy method
+- Improved ajax request settings [#27](http://github.com/rstaib/jquery-bootgrid/issues/27)
 
 ### Bug Fixes
 - ...